+++ frontend/rust-lib/flowy-chat/src/chat.rs
@@ -2,7 +2,7 @@ use crate::chat_manager::ChatUserService;
 use crate::entities::{
   ChatMessageErrorPB, ChatMessageListPB, ChatMessagePB, RepeatedRelatedQuestionPB,
 };
-use crate::middleware::chat_service_mw::ChatService;
+use crate::middleware::chat_service_mw::ChatServiceMiddleware;
 use crate::notification::{send_notification, ChatNotification};
 use crate::persistence::{insert_chat_messages, select_chat_messages, ChatMessageTable};
 use allo_isolate::Isolate;
@@ -11,6 +11,7 @@ use flowy_error::{FlowyError, FlowyResult};
 use flowy_sqlite::DBConnection;
 use futures::{SinkExt, StreamExt};
 use lib_infra::isolate_stream::IsolateSink;
+use std::path::PathBuf;
 use std::sync::atomic::{AtomicBool, AtomicI64};
 use std::sync::Arc;
 use tokio::sync::{Mutex, RwLock};
@@ -26,7 +27,7 @@ pub struct Chat {
   chat_id: String,
   uid: i64,
   user_service: Arc<dyn ChatUserService>,
-  chat_service: Arc<ChatService>,
+  chat_service: Arc<ChatServiceMiddleware>,
   prev_message_state: Arc<RwLock<PrevMessageState>>,
   latest_message_id: Arc<AtomicI64>,
   stop_stream: Arc<AtomicBool>,
@@ -38,7 +39,7 @@ impl Chat {
     uid: i64,
     chat_id: String,
     user_service: Arc<dyn ChatUserService>,
-    chat_service: Arc<ChatService>,
+    chat_service: Arc<ChatServiceMiddleware>,
   ) -> Chat {
     Chat {
       uid,
@@ -434,6 +435,33 @@ impl Chat {
       .collect::<Vec<_>>();
 
     Ok(messages)
+  }
+
+  #[instrument(level = "debug", skip_all, err)]
+  pub async fn index_file(&self, file_path: PathBuf) -> FlowyResult<()> {
+    if !file_path.exists() {
+      return Err(
+        FlowyError::record_not_found().with_context(format!("{:?} not exist", file_path)),
+      );
+    }
+
+    if !file_path.is_file() {
+      return Err(
+        FlowyError::invalid_data().with_context(format!("{:?} is not a file ", file_path)),
+      );
+    }
+
+    trace!(
+      "[Chat] index file: chat_id={}, file_path={:?}",
+      self.chat_id,
+      file_path
+    );
+    self
+      .chat_service
+      .index_file(&self.user_service.workspace_id()?, file_path, &self.chat_id)
+      .await?;
+
+    Ok(())
   }
 }