+++ frontend/rust-lib/flowy-search/tests/tantivy_test.rs
@@ -47,7 +47,7 @@ fn search_folder_test() {
 
   for (_score, doc_address) in top_docs {
     // Retrieve the actual content of documents given its `doc_address`.
-    let retrieved_doc = searcher.doc(doc_address).unwrap();
-    println!("{}", schema.to_json(&retrieved_doc));
+    let retrieved_doc: TantivyDocument = searcher.doc(doc_address).unwrap();
+    println!("{}", retrieved_doc.to_json(&schema));
   }
 }